requirements.txt
Django>=6.0.2
python-dotenv>=1.0.0
requests>=2.31.0
